refactor: [Step 5] 패키지 구분
