+++ src/main/java/racing/Race.java
@@ -1,6 +1,11 @@
 package racing;
 
-import racing.strategy.MoveStrategy;
+import racing.domain.Car;
+import racing.domain.Cars;
+import racing.domain.TryCount;
+import racing.view.InputView;
+import racing.view.ResultView;
+import racing.domain.strategy.MoveStrategy;
 
 import java.util.ArrayList;
 import java.util.List;
