+++ src/main/java/racing/domain/TryCount.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public class TryCount {
     private int tryCount;
