+++ src/main/java/racing/util/StringUtil.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.util;
 
 import java.util.Arrays;
 import java.util.List;
