+++ src/test/java/racing/CarsTest.java
@@ -2,7 +2,10 @@ package racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racing.strategy.MoveStrategy;
+import racing.domain.Car;
+import racing.domain.Cars;
+import racing.domain.Winners;
+import racing.domain.strategy.MoveStrategy;
 
 import java.util.List;
