+++ src/test/java/racing/NameTest.java
@@ -2,9 +2,10 @@ package racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.Name;
 
 import static org.assertj.core.api.Assertions.*;
-import static racing.Name.MAX_CAR_NAME_LENGTH;
+import static racing.domain.Name.MAX_CAR_NAME_LENGTH;
 
 class NameTest {
