+++ src/test/java/racing/StringUtilTest.java
@@ -3,6 +3,7 @@ package racing;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import racing.util.StringUtil;
 
 import java.util.List;
 import java.util.stream.Stream;
