+++ src/test/java/racing/TryCountTest.java
@@ -2,6 +2,7 @@ package racing;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.TryCount;
 
 import static org.assertj.core.api.Assertions.*;
